requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
praw==7.7.0
prawcore==2.3.0
requests==2.31.0
requests-toolbelt==1.0.0
async-timeout==5.0.1
redis==5.2.1
pymongo==4.6.1
python-dotenv==1.1.1
